[General] Changed code (forgor to commit before switching windows💀) #2
yess

+++ Frost/src/Frost/Platform/Vulkan/Buffers/VulkanIndexBuffer.cpp
@@ -6,7 +6,7 @@
 
 namespace Frost
 {
-	VulkanIndexBuffer::VulkanIndexBuffer(void* indicies, uint32_t count)
+	VulkanIndexBuffer::VulkanIndexBuffer(void* indicies, uint64_t count)
 		: m_BufferSize(count)
 	{
